更新生成数据集的代码，按照心拍作训练

+++ util/make_dataset.py
@@ -16,13 +16,10 @@ import csv
 import numpy as np
 import pandas as pd
 
-# import wfdb
-from pydub import AudioSegment
-from util.BEATs_def import get_wav_data
-
-
 # ========================/ functions define /========================== #
 # make dictionary
+
+
 def mkdir(path):
     folder = os.path.exists(path)
     # judge wether make dir or not
@@ -84,10 +81,8 @@ def copy_wav_file(src_path, folder_path, patient_id_list, mur, position):
         for pos in position:
             target_dir = folder_path + "\\" + mur + "\\" + patient_id + "\\"
             os.makedirs(target_dir, exist_ok=True)
-
             wavname = src_path + "\\" + patient_id + pos + ".wav"
             tsvname = src_path + "\\" + patient_id + pos + ".tsv"
-
             if os.path.exists(wavname):
                 shutil.copy(wavname, target_dir + "\\")
                 count += 1
@@ -137,7 +132,6 @@ def period_div(
                 dir_path = path + mur + patient_id + "\\" + patient_id + pos
                 tsv_path = dir_path + ".tsv"
                 wav_path = dir_path + ".wav"
-
                 index = id_data.index(patient_id)
                 wav_location = pos[1:]  # 听诊区域
                 locations = Murmur_locations[index].split("+")  # 有杂音的区域
@@ -155,7 +149,6 @@ def period_div(
                         Diastolic_murmur = "Absent"
                     else:
                         Diastolic_murmur = "Present"
-
                 # 此听诊区没有杂音
                 else:
                     Systolic_murmur = "Absent"
@@ -163,6 +156,7 @@ def period_div(
                     Systolic_state = "nan"
                     Diastolic_state = "nan"
                 if os.path.exists(tsv_path):
+                    # 切割数据
                     state_div(
                         tsv_path,
                         wav_path,
@@ -189,39 +183,40 @@ def state_div(
     index_file = index_load(tsvname)
     recording, fs = librosa.load(wavname, sr=4000)
     num = 0
-    start_index2 = 0
-    end_index2 = 0
-    start_index4 = 0
-    end_index4 = 0
+    # start_index1 = 0
+    # end_index1 = 0
+    # start_index2 = 0
+    # end_index2 = 0
 
     for i in range(index_file.shape[0] - 3):
-        if index_file[i][2] == "2" and index_file[i + 2][2] == "4":
-            start_index2 = float(index_file[i][0]) * fs
-            end_index2 = float(index_file[i][1]) * fs
-            start_index4 = float(index_file[i + 2][0]) * fs
-            end_index4 = float(index_file[i + 2][1]) * fs
+        if index_file[i][2] == "1" and index_file[i + 3][2] == "4":
+            start_index1 = float(index_file[i][0]) * fs
+            end_index1 = float(index_file[i+3][1]) * fs
+            start_index2 = float(index_file[i + 2][0]) * fs
+            end_index2 = float(index_file[i + 2][1]) * fs
             num = num + 1
             #  解决出现_0.wav的问题
-            print(start_index2, end_index2, start_index4, end_index4)
+            print(start_index1, end_index1, start_index2, end_index2)
             print("=============================================")
             print("wav name: " + wavname)
-            buff2 = recording[int(start_index2) : int(end_index2)]  # 字符串索引切割
-            buff4 = recording[int(start_index4) : int(end_index4)]  # 字符串索引切割
-            print("buff2 len: " + str(len(buff2)), "buff4 len: " + str(len(buff4)))
+            buff1 = recording[int(start_index1): int(end_index1)]  # 字符串索引切割
+            buff2 = recording[int(start_index2): int(end_index2)]  # 字符串索引切割
+            print("buff1 len: " + str(len(buff1)),
+                  "buff2 len: " + str(len(buff2)))
             soundfile.write(
                 state_path
                 + "{}_{}_{}_{}_{}.wav".format(
-                    index, "Systolic", num, Systolic_murmur, Systolic_state
+                    index, "s1+Systolic+s2+Diastolic", num, Systolic_murmur, Systolic_state
                 ),
-                buff2,
+                buff1,
                 fs,
             )
             soundfile.write(
                 state_path
                 + "{}_{}_{}_{}_{}.wav".format(
-                    index, "Diastolic", num, Diastolic_murmur, Diastolic_state
+                    index, "s2+Diastolic", num, Diastolic_murmur, Diastolic_state
                 ),
-                buff4,
+                buff2,
                 fs,
             )
 
@@ -235,10 +230,10 @@ def get_patientid(csv_path):
         return id
 
 
-# copy absent data to folder
+# copy data to folder
 def copy_states_data(folder, patient_id, murmur, type):
     din_path = folder + type + murmur
-    if os.path.exists(din_path):
+    if not os.path.exists(din_path):
         os.makedirs(din_path)
     for id in patient_id:
         dir_path = folder + murmur + id
@@ -247,8 +242,10 @@ def copy_states_data(folder, patient_id, murmur, type):
             for subdir in dir:
                 subdir_path = os.path.join(root, subdir)
                 print(subdir_path)
-                # if os.path.exists(dir_path):
-                shutil.copytree(subdir_path, din_path + subdir)
+                if os.path.exists(dir_path):
+                    shutil.copytree(subdir_path, din_path + subdir)
+                else:
+                    print("dir not exist")
 
 
 # ==================================================================== #
@@ -283,13 +280,13 @@ Diastolic_murmur_timing_path = (
     r"D:\Shilong\murmur\03_circor_states\Diastolic_murmur_timing.csv"
 )
 
-pd.DataFrame(Murmur_locations).to_csv(Murmur_locations_path, index=False, header=False)
-pd.DataFrame(Systolic_murmur_timing).to_csv(
-    Systolic_murmur_timing_path, index=False, header=False
-)
-pd.DataFrame(Diastolic_murmur_timing).to_csv(
-    Diastolic_murmur_timing_path, index=False, header=False
-)
+# pd.DataFrame(Murmur_locations).to_csv(Murmur_locations_path, index=False, header=False)
+# pd.DataFrame(Systolic_murmur_timing).to_csv(
+#     Systolic_murmur_timing_path, index=False, header=False
+# )
+# pd.DataFrame(Diastolic_murmur_timing).to_csv(
+#     Diastolic_murmur_timing_path, index=False, header=False
+# )
 
 # init aptient id list for absent present and unknown
 absent_patient_id = list()
@@ -314,17 +311,14 @@ for id in present_id:
 positoin = ["_AV", "_MV", "_PV", "_TV"]
 murmur = ["Absent\\", "Present\\"]
 period = ["s1", "systolic", "s2", "diastolic"]
-folder_path = r"D:\Shilong\murmur\try\\"
-
 src_path = r"D:\Shilong\murmur\dataset_all\training_data"
-# # make dir and copy files for Present/Absent patients
+folder_path = r"D:\Shilong\murmur\01_dataset\02_period\\"
+# 将wav文件和tsv文件copy到目标文件夹
 copy_wav_file(src_path, folder_path, absent_patient_id, "Absent", positoin)
 copy_wav_file(src_path, folder_path, present_patient_id, "Present", positoin)
-
-# make dir for each position
 # D:\Shilong\murmur\LM_wav_dataset
 src_path = r"D:\Shilong\murmur\dataset_all\training_data"
-
+# 创建每个wav文件的文件夹
 for mur in murmur:
     dir_path = folder_path + mur
     for patient_id in absent_patient_id:
@@ -333,6 +327,8 @@ for mur in murmur:
         pos_dir_make(dir_path, patient_id, positoin)
 
 # 切数据，命名格式为：id+pos+state+num
+# 对数据打标签
+# absent
 period_div(
     folder_path,
     murmur,
@@ -343,6 +339,7 @@ period_div(
     Systolic_murmur_timing,
     Diastolic_murmur_timing,
 )
+# present
 period_div(
     folder_path,
     murmur,
@@ -364,12 +361,13 @@ present_test_id_path = r"D:\Shilong\murmur\03_circor_states\present_test_id.csv"
 # present_train_id=random.sample(present_patient_id,int(len(present_patient_id)*0.8))
 # absent_test_id=list(set(absent_patient_id)-set(absent_train_id))
 # present_test_id=list(set(present_patient_id)-set(present_train_id))
+# 读取训练集和测试集id划分
 absent_train_id = csv_reader_cl(absent_train_id_path, 0)
 absent_test_id = csv_reader_cl(absent_test_id_path, 0)
 present_train_id = csv_reader_cl(present_train_id_path, 0)
 present_test_id = csv_reader_cl(present_test_id_path, 0)
 # 将训练集和测试集文件分别copy到train和test文件夹
-folder = r"D:\Shilong\murmur\03_circor_states"
+folder = r"D:\Shilong\murmur\01_dataset\02_period"
 copy_states_data(folder, absent_train_id, "\\Absent\\", "\\train")
 copy_states_data(folder, present_train_id, "\\Present\\", "\\train")
 copy_states_data(folder, absent_test_id, "\\Absent\\", "\\test")
@@ -380,94 +378,3 @@ copy_states_data(folder, present_test_id, "\\Present\\", "\\test")
 # pd.DataFrame(present_train_id).to_csv(present_train_id_path, index=False, header=False)
 # pd.DataFrame(absent_test_id).to_csv(absent_test_id_path, index=False, header=False)
 # pd.DataFrame(present_test_id).to_csv(present_test_id_path, index=False, header=False)
-
-
-# ========================/ parameteres define /========================== #
-murmur_positoin = ["_AV", "_MV", "_PV", "_TV"]
-murmur_ap = ["Absent\\", "Present\\"]
-period = ["Systolic", "Diastolic"]
-
-# ========================/ file path /========================== #
-# get absent / present patient_id
-id_data_path = r"D:\Shilong\murmur\03_circor_states\id_data.csv"
-absent_csv_path = r"D:\Shilong\murmur\03_circor_states\absent_id.csv"
-present_csv_path = r"D:\Shilong\murmur\03_circor_states\present_id.csv"
-Diastolic_murmur_timing_path = (
-    r"D:\Shilong\murmur\03_circor_states\Diastolic_murmur_timing.csv"
-)
-Systolic_murmur_timing_path = (
-    r"D:\Shilong\murmur\03_circor_states\Systolic_murmur_timing.csv"
-)
-Murmur_locations_path = r"D:\Shilong\murmur\03_circor_states\Murmur_locations.csv"
-
-
-filepath = r"D:\Shilong\murmur\03_circor_states"
-absent_train_path = r"D:\Shilong\murmur\03_circor_states\trainset\absent"
-absent_test_path = r"D:\Shilong\murmur\03_circor_states\testset\absent"
-present_train_path = r"D:\Shilong\murmur\03_circor_states\trainset\present"
-present_test_path = r"D:\Shilong\murmur\03_circor_states\testset\present"
-present_train_path_8 = r"D:\Shilong\murmur\03_circor_states\trainset\time_stretch0.8"
-present_train_path_12 = r"D:\Shilong\murmur\03_circor_states\trainset\time_stretch1.2"
-
-folder = r"D:\Shilong\murmur\03_circor_statest"
-npy_path = r"D:\Shilong\murmur\03_circor_states\npyFile"
-
-path = r"D:\Shilong\murmur\03_circor_states\csv"
-train_path = r"D:\Shilong\murmur\03_circor_states\train_csv"
-test_path = r"D:\Shilong\murmur\03_circor_states\test_csv"
-# ========================/ get lists /========================== #
-id_data = get_patientid(id_data_path)
-absent_patient_id = get_patientid(absent_csv_path)
-present_patient_id = get_patientid(present_csv_path)
-Diastolic_murmur_timing = get_patientid(Diastolic_murmur_timing_path)
-Systolic_murmur_timing = get_patientid(Systolic_murmur_timing_path)
-Murmur_locations = get_patientid(Murmur_locations_path)
-
-absent_train_csv_path = r"D:\Shilong\murmur\03_circor_states\train_csv"
-absent_test_csv_path = r"D:\Shilong\murmur\03_circor_states\test_csv"
-present_train_csv_path = r"D:\Shilong\murmur\03_circor_states\train_csv"
-present_test_csv_path = r"D:\Shilong\murmur\03_circor_states\test_csv"
-
-filepath = r"D:\Shilong\murmur\03_circor_states"
-absent_train_path = r"D:\Shilong\murmur\03_circor_states\trainset\absent"
-absent_test_path = r"D:\Shilong\murmur\03_circor_states\testset\absent"
-present_train_path = r"D:\Shilong\murmur\03_circor_states\trainset\present"
-present_test_path = r"D:\Shilong\murmur\03_circor_states\testset\present"
-present_train_path_8 = r"D:\Shilong\murmur\03_circor_states\trainset\time_stretch0.8"
-present_train_path_12 = r"D:\Shilong\murmur\03_circor_states\trainset\time_stretch1.2"
-npy_path_padded = r"D:\Shilong\murmur\03_circor_states\npyFile_padded"
-
-
-absent_train_features, absent_train_label = get_wav_data(
-    absent_train_path, absent_train_csv_path
-)  # absent
-absent_test_features, absent_test_label = get_wav_data(
-    absent_test_path, absent_test_csv_path
-)  # absent
-present_train_features, present_train_label = get_wav_data(
-    present_train_path, present_train_csv_path
-)  # present
-present_test_features, present_test_label = get_wav_data(
-    present_test_path, present_test_csv_path
-)  # present
-present_train_features_8, present_train_label_8 = get_wav_data(
-    present_train_path_8
-)  # present
-present_train_features_12, present_train_label_12 = get_wav_data(
-    present_train_path_12
-)  # present
-# # # ========================/ save as npy file /========================== #
-np.save(npy_path_padded + r"\absent_train_features.npy", absent_train_features)
-np.save(npy_path_padded + r"\absent_test_features.npy", absent_test_features)
-np.save(npy_path_padded + r"\present_train_features.npy", present_train_features)
-np.save(npy_path_padded + r"\present_test_features.npy", present_test_features)
-
-np.save(npy_path_padded + r"\absent_train_label.npy", absent_train_label)
-np.save(npy_path_padded + r"\absent_test_label.npy", absent_test_label)
-np.save(npy_path_padded + r"\present_train_label.npy", present_train_label)
-np.save(npy_path_padded + r"\present_test_label.npy", present_test_label)
-
-np.save(npy_path_padded + r"\present_train_features_8.npy", present_train_features_8)
-np.save(npy_path_padded + r"\present_train_features_12.npy", present_train_features_12)
-np.save(npy_path_padded + r"\present_train_label_8.npy", present_train_label_8)
-np.save(npy_path_padded + r"\present_train_label_12.npy", present_train_label_12)
